change:  make a dynamic element based role on siswa module

+++ routes/web.php
@@ -19,11 +19,11 @@ Route::middleware('auth')->group(function () {
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 });
 
-Route::middleware(['auth', 'role:admin'])->prefix('master')->name('master.')->group(function () {
+Route::middleware(['auth'])->prefix('master')->name('master.')->group(function () {
     Route::get('/siswa', [MasterSiswaController::class, 'index'])->name('siswa');
-    Route::post('/siswa/store', [MasterSiswaController::class, 'store_siswa'])->name('store.siswa');
-    Route::put('/siswa/update/{user_siswa_id}', [MasterSiswaController::class, 'update_siswa'])->name('update.siswa');
-    Route::delete('/siswa/{user_id}', [MasterSiswaController::class, 'destroy_siswa'])->name('destroy.siswa');
+    Route::post('/siswa/store', [MasterSiswaController::class, 'store_siswa'])->name('store.siswa')->middleware('role:admin');
+    Route::put('/siswa/update/{user_siswa_id}', [MasterSiswaController::class, 'update_siswa'])->name('update.siswa')->middleware('role:admin');
+    Route::delete('/siswa/{user_id}', [MasterSiswaController::class, 'destroy_siswa'])->name('destroy.siswa')->middleware('role:admin');
 });
 
 require __DIR__ . '/auth.php';
